crawlers/requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.9.3
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.1.1
decorator==4.4.2
idna==2.10
pycparser==2.20
python-dotenv==0.14.0
python-telegram-bot==13.0
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
telebot==0.0.3
tornado==6.0.4
tzlocal==2.1
urllib3==1.25.10
